backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
httpx==0.28.1
ytmusicapi==1.8.2
yt-dlp==2024.12.13
redis==5.2.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
loguru==0.7.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.4
